test: add coverage config
